moved two main functions out to lib file
